+++ src/strategies/curve/strategy-curve-scrv-v3_2.sol
@@ -47,7 +47,7 @@ contract StrategyCurveSCRVv3_2 is StrategyCurveBase {
         public
         override
         view
-        returns (address, uint256)
+        returns (address, uint8)
     {
         uint256[] memory balances = new uint256[](4);
         balances[0] = ICurveFi_4(curve).balances(0); // DAI
@@ -109,7 +109,7 @@ contract StrategyCurveSCRVv3_2 is StrategyCurveBase {
         //      if so, a new strategy will be deployed.
 
         // stablecoin we want to convert to
-        (address to, uint256 toIndex) = getMostPremium();
+        (address to, uint8 toIndex) = getMostPremium();
 
         // Collects crv tokens
         // Don't bother voting in v1
